Adding a new Config Class called ServerConfig for Server specific properties.
Adding 2 new properties to the newly created ServerConfig, and mapping them for ENV override using the ServerMappedConfigSourceInitializer. maxMessageSizeBytes and port.

Replacing the hardcoded values on the codebase with the new configurable properties from ServerConfig.

Added the new configuration to the ConfigInjectionModule

Created UT for testing the Config validations and restrictions.

Added Test case for ConfigInjectionModule to retrieve the newly config object.

Signed-off-by: Alfredo Gutierrez <alfredo@swirldslabs.com>