feat: add support for `PublishBlockStream` to simulator (#548)
This PR adds support for PublishBlockStream in the simulator, by introducing new wokring mode, called PublisherServerMode, where the application acts a server, which listents and awaits stream of block data. With the effort here I aim mainly to introduce the feature, where in future ones we will refine it, according to needs to be able to use it as testing driver.

Signed-off-by: georgi-l95 <glazarov95@gmail.com>